requirements.txt
requests
pytest
pytest-cov
mysql-connector-python
